requirements.txt
fastapi==0.115.14
httpx==0.27.2
passlib==1.7.4
pydantic==2.11.7
pydantic_settings==2.10.1
cryptography==43.0.1
PyJWT==2.10.1
pyotp==2.9.0
PyQt5==5.15.11
PyQt5_sip==12.17.0
pytest==8.2.2
python-dotenv==1.1.1
qrcode==7.4.2
Requests==2.32.4
setuptools==69.5.1
SQLAlchemy==2.0.41
stripe==12.2.0
uvicorn==0.35.0
prometheus-client==0.21.0
watchdog==6.0.0
pydantic[email]
email-validator
